Improved Navigation Implementation
Instead of a bunch of switch statemend and rewriting the entire list of menu buttons everytime adding THOUSANDS of lines of code, HighlightedCheck() now takes care of it. Added UpdateIndexValues() so IntraMenu index doesnt go outside the current Menu's buttons, same thing the default section of the switch statements did in the old navigation.

Reorganized DrawMenu so that the methods were in order.